% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainAgeShiftObj_methods.R
\name{default_slot}
\alias{default_slot}
\title{Change the default_slot}
\usage{
default_slot(obj, ...)
}
\arguments{
\item{obj}{A brainAgeShift object.}
}
\description{
Changes the default_slot, which is the one that will be used by predictAge.
to compute transcriptomic age.
}
\examples{
default_slot(obj) <- "norm_counts"
}
