% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainAgeShiftObj_methods.R
\name{do_frozenSVA}
\alias{do_frozenSVA}
\title{Apply frozen SVA}
\usage{
do_frozenSVA(obj, ...)
}
\arguments{
\item{obj}{A brainAgeShift object.}
}
\value{
A brainAgeShift object with the frozen_SVAed slot filled with the
normalized counts matrix.
}
\description{
Apply frozen SVA on the norm_counts slot. Predicts SVs based on the ones
computed in our training data, and regresses them out. Default slot is
changed to frozen_SVAed after running this method.
}
