% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example dataset}
\format{
example_counts (26189 gene counts x 74 samples). example_metadata
(6 columns, containing GSE, GSM, sample name, treatment, pheono and age).
}
\source{
DOI: 10.1126/science.adl2992
}
\usage{
data(example_data)
}
\description{
example_counts and example_metadata datasets to illustrate the use of the
package.
}
