% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainAgeShiftObj_methods.R
\name{get_signGenes}
\alias{get_signGenes}
\title{Obtain genes significantly contributing to age shifts}
\usage{
get_signGenes(obj, ...)
}
\arguments{
\item{obj}{A brainAgeShift object with a not-empty stats slot.}

\item{alpha_comparisons}{The level of significance of the comparisons.}

\item{alpha_genes}{The level of significance of the genes.}

\item{n_perms}{The number of permutations for running the permutation test.}

\item{adjust_method}{The p-value adjustment method for the genes' p-values.}

\item{sort_genes}{If genes should be sorted according to weighted
differences in each one of the sign_genes dataframes. If true, if the log2FC
of the comparison is positive genes will be sorted in a decreasing fashion,
and the other way around.}
}
\value{
A brainAgeShift object with the sign_genes slot filled with a list
where each item consists on a dataframe of the significant genes
(p_adj <= alpha_genes) for each one of the significant comparisons
(p_adj <= alpha_comparisons). This dataframe contains the following columns:
geneID, difference, coefficient, weighted_diff, p_value and p_adj.
}
\description{
For each one of the significant comparisons, determine which are the genes
that are significantly contributing to such shift. For that, the weighted
mean difference between each group of each significant comparison is computed
for each gene, and is compared to a null distribution obtained through random
permutation of the group labels, obtaining a p-value for each gene, which
is later adjusted.
}
