% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainAgeShiftObj_methods.R
\name{normalizeCounts}
\alias{normalizeCounts}
\alias{normalizeCounts.brainAgeShiftObj}
\title{Normalize brainAgeShift counts}
\usage{
normalizeCounts(obj, ...)

\method{normalizeCounts}{brainAgeShiftObj}(obj, useTrainMeans = F)
}
\arguments{
\item{obj}{A \code{brainAgeShift} object.}

\item{useTrainMeans}{Logical. If \code{TRUE}, uses the quantile means from the
training data for nomralization, ensuring distributions are closer to the
training set. If \code{FALSE}, computes quantile means directly from the input
data.}
}
\value{
A \code{brainAgeShift} object with the \code{norm_counts} slot filled with the
normalized counts matrix.
}
\description{
Normalizes the counts in a \code{brainAgeShift} object to prepare them for age
prediction. The normalization process includes filtering genes to retain only
those used in training, applying log2 transformation (\code{log2(counts + 1)}),
and performing quantile normalization. After normalization, the
\code{default_slot} of the object is updated to \code{"norm_counts"}.
}
\examples{
# Normalize counts using computed quantile means
obj <- normalizeCounts(obj, useTrainMeans = FALSE)

# Normalize counts using training quantile means
obj <- normalizeCounts(obj, useTrainMeans = TRUE)
}
