% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainAgeShiftObj_methods.R
\name{predictAge}
\alias{predictAge}
\title{Predict ages
Uses the brain transcriptomic clock to predict transcriptomic ages, based on
the default slot ("counts", "norm_counts", "frozen_SVAed")..}
\usage{
predictAge(obj, ...)
}
\arguments{
\item{obj}{A brainAgeShift object.}
}
\value{
A brainAgeShift object with a predicted_age column added to the
metadata slot dataframe.
}
\description{
Predict ages
Uses the brain transcriptomic clock to predict transcriptomic ages, based on
the default slot ("counts", "norm_counts", "frozen_SVAed")..
}
