% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainAgeShiftObj_def.R
\name{create_brainAgeShiftObj}
\alias{create_brainAgeShiftObj}
\title{Create a brainAgeShift object.}
\usage{
create_brainAgeShiftObj(
  counts,
  metadata = NULL,
  variable = NULL,
  comparisons = NULL
)
}
\arguments{
\item{counts}{A matrix of gene counts. Genes in rows, samples in columns. Row
names must be ENSEMBL IDs}

\item{metadata}{A data.frame with at least the variable of interest for
comparing transcriptomic ages. Row names must match column names of counts
matrix. Defaults to \code{NULL}. If not stated, the resulting object can't be used
beyond age prediction.}

\item{variable}{The column in the metadata dataframe where the comparisons
are stored. Defaults to \code{NULL}. If not stated, the resulting object can't be
used beyond age prediction.}

\item{comparisons}{Either a vector of two factors, or a list of multiple
two-factor vectors. First item in each comparison will be initial or base
state during the difference computation, second the final or altered state.
Defaults to \code{NULL}. If not stated, the resulting object can't be used beyond
age prediction.}
}
\value{
A new brainAgeShift object initialized with the required slots and
default values.
}
\description{
Creates a \code{brainAgeShift} object to perform brain transcriptomic age
predictions and subsequent statistical analyses to detect transcriptomic
age-shifting factors.
}
