% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainAgeShiftObj_methods.R
\name{do_frozenSVA}
\alias{do_frozenSVA}
\alias{do_frozenSVA.brainAgeShiftObj}
\title{Apply frozen SVA}
\usage{
do_frozenSVA(obj, ...)

\method{do_frozenSVA}{brainAgeShiftObj}(obj)
}
\arguments{
\item{obj}{A \code{brainAgeShift} object.}
}
\value{
A \code{brainAgeShift} object with the \code{frozen_SVAed} slot filled with the
SVAed normalized counts matrix.
}
\description{
Applies frozen SVA on the \code{norm_counts} slot of a \code{brainAgeShift} object.\
Predicts SVs based on the ones computed in our training data, and regresses
them out. Default slot is changed to \code{frozen_SVAed} after running this
method.
}
