% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainAgeShiftObj_methods.R
\name{do_signTest}
\alias{do_signTest}
\title{Compute comparison significance.}
\usage{
do_signTest(obj, ...)
}
\arguments{
\item{obj}{A brainAgeShift object with a not-empty metadata slot, a variable
and at least one comparison.}

\item{adjust_method}{The p-value adjustment method.}
}
\value{
A brainAgeShift object with the stats slot filled with a dataframe
containing the results of the statistical analyses for each comparison.
}
\description{
Compute p-value (t-test), adjusted p-value, log2 fold change and difference
of each one of the specified comparisons.
}
