% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example dataset}
\format{
example_counts (63677 transcript counts x 9 samples). example_metadata
(8 columns, containing specimenID, RIN, individualID, organ, sex, race, age and group).
}
\source{
DOI: 10.1126/science.adl2992
}
\usage{
data(example_data)
}
\description{
example_counts and example_metadata datasets to illustrate the use of the
package.
}
