% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainAgeShiftObj_methods.R
\name{get_signGenes}
\alias{get_signGenes}
\alias{get_signGenes.brainAgeShiftObj}
\title{Identify genes significantly contributing to age shifts}
\usage{
get_signGenes(obj, ...)

\method{get_signGenes}{brainAgeShiftObj}(
  obj,
  alpha_comparisons = 0.05,
  alpha_genes = 0.05,
  n_perms = 1000,
  adjust_method = "BH",
  sort_genes = T
)
}
\arguments{
\item{obj}{A \code{brainAgeShift} object with a non-empty \code{stats} slot.}

\item{alpha_comparisons}{Numeric. The significance threshold for selecting
significant comparisons (adjusted p-value <= \code{alpha_comparisons}). Defaults
to \code{0.05}.}

\item{alpha_genes}{Numeric. The significance threshold for selecting
significant genes (adjusted p-value <= \code{alpha_genes}). Defaults to \code{0.05}.}

\item{n_perms}{Integer. The number of permutations to perform for the
permutation test. Defaults to \code{1000}.}

\item{adjust_method}{Character. The p-value adjustment method to use for
genes' p-values (e.g. \code{"BH"} for Benjamini-Hochberg or \code{"bonferroni"}).
Defaults to \code{"BH"}.}

\item{sort_genes}{Logical. If \code{TRUE}, genes in each \code{sign_genes} dataframe
will be sorted by weighted differences. Sorting is performed such that
positive log2 fold change in the comparison implies decreasing order of
genes' weighted differences, and viceversa. Defaults to \code{TRUE}.}
}
\value{
A \code{brainAgeShift} object with the \code{sign_genes} slot updated. The
\code{sign_genes} slot contains a list where each element corresponds to a
significant comparison and is a dataframe of significant genes. The
dataframes include the following columns:
\itemize{
\item \strong{geneID}: the ENSEMBL ID of the gene.
\item \strong{difference}: the raw mean difference between groups.
\item \strong{coefficient}: the brain clock coefficient.
\item \strong{weighted_diff}: the weighted mean difference
\item \strong{p_value}: the p-value from the permutation test.
\item \strong{p_adj}: the adjusted p-value.
}
}
\description{
For each significant comparison in the \code{stats} slot of a \code{brainAgeShift}
object, this function identifies the genes that significantly contribute to
observed shifts. The weighted mean difference between comparison groups is
calculated for each gene and compared against a null distribution generated
through random permutations of group labels. P-values are obtained for each
gene and adjusted for multiple comparisons.
}
\examples{
# Assuming obj is a brainAgeShift object with non-empty stats slot:
obj <- get_signGenes(obj,
                     alpha_comparisons = 0.05, 
                     alpha_genes = 0.01,
                     n_perms = 1000, 
                     adjust_method = "BH",
                     sort_genes = TRUE)
}
