% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainAgeShiftObj_methods.R
\name{normalizeCounts}
\alias{normalizeCounts}
\title{Normalize brainAgeShift counts}
\usage{
normalizeCounts(obj, ...)
}
\arguments{
\item{obj}{A brainAgeShift object.}

\item{useTrainMeans}{Instead of computing the quantile means on the
introduced data, uses the quantiles from our training data, to make
distributions closer. Either TRUE or FALSE.}
}
\value{
A brainAgeShift object with the norm_counts slot filled with the
normalized counts matrix.
}
\description{
Normalize the counts of the brainAgeShift object to make them suitable
to predict ages with it. First it filters the genes to keep only training
ones, then computes log2(counts + 1), and after applies quantile
normalization. Default slot is changed to norm_counts after running
this method.
}
