% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainAgeShiftObj_methods.R
\name{predictAge}
\alias{predictAge}
\alias{predictAge.brainAgeShiftObj}
\title{Predict ages}
\usage{
predictAge(obj, ...)

\method{predictAge}{brainAgeShiftObj}(obj)
}
\arguments{
\item{obj}{A \code{brainAgeShift} object.}
}
\value{
A \code{brainAgeShift} object with a \code{predicted_age} column added to the
metadata slot's dataframe.
}
\description{
Predicts transcriptomic ages using the brain transcriptomic clock. The method
operates based on the default slot, which must be one of \code{counts} or
\code{norm_counts}.
}
\examples{
obj <- predictAge(obj)
}
